% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1 update lotto data.R
\name{data_update}
\alias{data_update}
\title{generate lottodata}
\usage{
data_update()
}
\value{
lotto data
}
\description{
generate lottodata
}
\examples{
devtools::install_github("qkdrk777777/DUcj")

data<-data_update()
}
