% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_multi.R
\name{multi_multi}
\alias{multi_multi}
\title{Multi Multi}
\usage{
multi_multi(draw_number = 10)
}
\arguments{
\item{draw_number}{Default is draw_numbers=10 - the maximum draws in multi multi game, and the possible values of draw_numbers are integers from range 1-10 like in the real game.}
}
\description{
Draws as many numbers as specified in draw_number parameter from 80. Default is draw_numbers=10 - the maximum draws in multi multi game, and the possible values of draw_numbers are integers from range 1-10 like in the real game.
}
