% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate lottodata.R
\name{numdata}
\alias{numdata}
\title{generate lottodata}
\usage{
numdata(a, b, first = 1, last, css = ".num_box", encoding = "UTF-8",
  num = list(), output = list(), k = 1)
}
\arguments{
\item{css}{= Nodes to select. Supply one of css or xpath depending on whether you want to use a css or xpath 1.0 selector.}

\item{t}{= url of list form}
}
\description{
generate lottodata
}
\examples{
a='https://search.naver.com/search.naver?sm=tab_drt&where=nexearch&query='
b='\%ED\%9A\%8C\%EB\%A1\%9C\%EB\%98\%90'
url<-'https://search.naver.com/search.naver?sm=tab_hty.top&where=nexearch&query=\%EB\%A1\%9C\%EB\%98\%90&oquery=\%EB\%A1\%9C\%EB\%98\%90&tqi=TmrKidpVuFdsssc0EvVssssssUd-075322'
line<-read_html(url,encoding="UTF-8")
p1<-html_nodes(line,css='._lotto-btn-current em')\%>\%html_text()
last=as.numeric(substr(p1,1,3))
(data<-numdata(a,b,last=last))
}
