% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbDSena.R
\name{ProbDSena}
\alias{ProbDSena}
\title{Dupla Sena lottery game probability}
\usage{
ProbDSena(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
Does the ticket have 50 tens, in which the bettor can? dial from 6 to 15 numbers,
consists of extracting 6 different numbers, you can win by hitting 3, 4, 5 or
6 numbers, with a differential, the player has two chances to win.
}
\details{
To choose 6 numbers, the bet is simple, where it offers the lowest odds
win: combination of 50 numbers taken from 6 to 6 = 50!/6!44! = 15,890,700
However, as there are two draws, the player has two chances out of approximately 16 million.
2/15,890,700 = 1/7945,350, to win any of the SENAS, the player has approximately 1 chance in 8 million.

x is the amount of numbers bet

y is the amount number of hits
}
\examples{
ProbDSena(15,5)
}
