% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbDSorte.R
\name{ProbDSorte}
\alias{ProbDSorte}
\title{Dia de Sorte lottery game probability}
\usage{
ProbDSorte(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
The player simply chooses from 7 to 15 numbers among the 31 available, they are extracted
7 numbers, you can win by hitting 7, 6, 5 or 4 numbers.
}
\details{
warning message if y (number of hits) is greater than 7, "Prob = 0, the maximum number
of hits is 7". However, if y <4, note that this probability does not exist.
Otherwise note that the player has won.

x is the amount of numbers bet

y is the amount number of hits
}
\examples{
ProbDSorte(7,6)
}
