% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbLFacil.R
\name{ProbLFacil}
\alias{ProbLFacil}
\title{Lotofacil lottery game probability}
\usage{
ProbLFacil(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
The player simply chooses and dials between 15 and 20 numbers, among the 25 available,
the draw consists of extracting 15 different numbers, in the universe from 1 to 25. In this game,
the prizes are awarded to the winners of 11, 12, 13, 14 or 15 numbers.
}
\details{
of hits is 15". However if y < 11, notice that there is no such probability.
Otherwise, notice that the bettor has won.

x is the amount of numbers bet

y is the amount number of hits
}
\examples{
ProbLFacil(15,11)
}
