% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbLMania.R
\name{ProbLMania}
\alias{ProbLMania}
\title{LotoMania lottery game probability}
\usage{
ProbLMania(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
The player only has to choose 50 numbers out of 100 available,
the draw consists of extracting 20 different numbers, between 00 and 99.
In this game, winners of 20, 19, 18, 17, 16, 15 or those who do not match
any number win prizes.
}
\details{
x the amount of numbers bet

y the amount number of hits
}
\examples{
ProbLMania(50,20)
}
