% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbMSena.R
\name{ProbMSena}
\alias{ProbMSena}
\title{Mega Sena lottery game probability}
\usage{
ProbMSena(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
The bettor simply chooses and dials 6 to 15 numbers from 60 available,
the draw consists of the extraction of 6 different numbers, in the universe from 1 to 60.
In that game, winners of 4,5 or 6 numbers win prizes.
}
\details{
Warning message if the y (number of hits) is greater than 6, "Prob = 0, the maximum number
of hits is 6". However if y <= 3, notice that there is no such probability.
Otherwise, notice that the bettor has won.

x is the amount of numbers bet

y is the amount number of hits
}
\examples{
ProbMSena(6,9)
}
