% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbQuina.R
\name{ProbQuina}
\alias{ProbQuina}
\title{Quina lottery game probability}
\usage{
ProbQuina(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
The player simply chooses and dials from 5 to 15 numbers among the 80 available,
the draw consists of extracting 5 different numbers, in the universe from 1 to 80.
In this game, the winners of 2(duque), 3(terno), 4(quadra) win
or 5(quina) numbers.
}
\details{
warning message if y (number of hits) is greater than 5, "Prob = 0, the maximum number
of hits is 5". However, if y <2, note that this probability does not exist.
Otherwise note that the player has won.

x is the amount of numbers bet

y is the number of hits
}
\examples{
ProbQuina(6,7)

}
