% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbTMania.R
\name{ProbTMania}
\alias{ProbTMania}
\title{Timemania lottery game probability}
\usage{
ProbTMania(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
number
}
\description{
The player chooses 10 numbers from the 80 available, 7 numbers are extracted,
then to compete for the prizes for hits of 7, 6, 5, 4, 3 and the team of heart.
}
\details{
Warning message if y (number of hits) is greater than 7, "Prob = 0, the maximum number
of hits is 7". However, if y <3, note that this probability does not exist.
Otherwise note that the player has won.

x is the amount of numbers bet

y is the amount number of hits
}
\examples{
ProbTMania(10,9)
}
