% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6 sampling2.R
\name{repsample}
\alias{repsample}
\title{sample2}
\usage{
repsample(n, count = 1, keep = NULL, drop = NULL, hold = NULL,
  up = NULL, header = 10, write = F)
}
\arguments{
\item{n}{= The number of times the sample is pulled}

\item{keep}{= The number to keep the probability of}

\item{drop}{=The number corresponding to 'drop' does not appear.}

\item{hold}{=This number appears unconditionally.}
}
\description{
sample2
}
\examples{
repsample(1000,up=1)
repsample(1000,header=10)
(k<-sort(num_per()[as.numeric(data[1,1:6])],dec=T))
}
