% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate lottodata.R
\name{numdata}
\alias{numdata}
\title{generate lottodata}
\usage{
numdata(a, b, first = 1, last, css = ".num_box", encoding = "UTF-8",
  num = list(), output = list(), k = 1)
}
\description{
generate lottodata
}
\examples{
a='https://search.naver.com/search.naver?sm=tab_drt&where=nexearch&query='
b='\%ED\%9A\%8C\%EB\%A1\%9C\%EB\%98\%90'
numdata(a,b,last=789)
}
